+++ src/packages/cli/.gitignore
@@ -19,6 +19,7 @@ pkg-build
 download-done
 prisma-client
 preinstall
+install
 
 prisma
 .envrc
