--- src/packages/cli/scripts/copy-runtime-dist.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-
-set -ex
-
-cp -R "$(./scripts/resolve.js @prisma/studio)/build/" dist/public
