+++ src/packages/client/src/__tests__/integration/happy/insensitive-postgresql-feature-flag/test.ts
@@ -228,6 +228,6 @@ test('insensitive-postgresql', async () => {
 
   prisma.$disconnect()
   await tearDownPostgres(SetupParams).catch((e) => {
-    console.log(e)
+    // console.log(e)
   })
 })
