+++ src/packages/engines/src/index.ts
@@ -3,3 +3,5 @@ import path from 'path'
 export function getEnginesPath() {
   return path.join(__dirname, '../')
 }
+
+export { enginesVersion } from '@prisma/engines-version'
