+++ src/packages/sdk/package.json
@@ -11,8 +11,6 @@
     "prepublishOnly": "pnpm run build && pnpm run test",
     "format": "prettier --write .",
     "lint": "eslint --fix --ext .js,.ts .",
-    "postinstall": "node scripts/download.js",
-    "download": "node scripts/updateTag.js && node scripts/download.js",
     "test": "jest"
   },
   "files": [
@@ -20,9 +18,6 @@
     "dist",
     "scripts"
   ],
-  "prisma": {
-    "version": "369b3694b7edb869fad14827a33ad3f3f49bbc20"
-  },
   "devDependencies": {
     "@types/jest": "26.0.14",
     "@types/node": "12.12.62",
@@ -45,6 +40,7 @@
   "dependencies": {
     "@prisma/debug": "workspace:*",
     "@prisma/engine-core": "workspace:*",
+    "@prisma/engines": "workspace:^0.0.1-e078aa75c40550d826bc35aeee4f45582fb4165e",
     "@prisma/fetch-engine": "workspace:*",
     "@prisma/generator-helper": "workspace:*",
     "@prisma/get-platform": "workspace:*",
